compilations

+++ ewm-service/src/main/java/ru/practicum/explore/server/category/service/CategoryServiceImpl.java
@@ -6,15 +6,15 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import ru.practicum.explore.server.category.dal.CategoryMapper;
+import ru.practicum.explore.server.category.dal.CategoryRepository;
 import ru.practicum.explore.server.category.dto.CategoryResponseDto;
 import ru.practicum.explore.server.category.dto.NewCategoryDto;
 import ru.practicum.explore.server.category.dto.UpdateCategoryDto;
-import ru.practicum.explore.server.category.dal.CategoryMapper;
-import ru.practicum.explore.server.category.dal.CategoryRepository;
 import ru.practicum.explore.server.category.model.Category;
+import ru.practicum.explore.server.event.repository.EventRepository;
 import ru.practicum.explore.server.exception.ConflictException;
 import ru.practicum.explore.server.exception.NotFoundException;
-
 
 import java.util.Collection;
 
@@ -23,6 +23,7 @@ import java.util.Collection;
 @Slf4j
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryResponseDto create(NewCategoryDto dto) {
@@ -74,7 +75,11 @@ public class CategoryServiceImpl implements CategoryService {
     public void delete(long id) {
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Категория с id = " + id + " не найдена"));
-        // когда появятся события — проверять отсутствие связей
+
+        if (eventRepository.existsByCategory_Id(id)) {
+            throw new ConflictException("Нельзя удалить категорию, к которой привязаны события");
+        }
+
         categoryRepository.deleteById(id);
         log.info("Category with id={} was deleted", id);
